resmgr: move/generate resmgr API/types to pkg/apis.
Move resmgr API and types to pkg/apis/resmgr/v1alpha1,
next to other generated APIs (currently config). This
allows annotating the types for better CRD validation
which is useful if we use them in our config CRDs.
